+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
  * Added support for passing through `:tag` metadata to the generated Python AST (#354)
+ * Added support for calling symbols as functions on maps and sets (#775)
+ * Added support for passing command line arguments to Basilisp (#???)
 
 ### Changed
  * Optimize calls to Python's `operator` module into their corresponding native operators (#754)
  * Allow vars to be callable to adhere to Clojure conventions (#767)
- * Support symbols as fns for sets and maps (#775)
 
 ### Fixed
  * Fix issue with `(count nil)` throwing an exception (#759)